requirements.txt
PyQt5>=5.15
mss>=9.0
requests>=2.31
Pillow>=10.0
numpy>=1.24
PyYAML>=6.0
pywin32>=306
